requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
protobuf==4.21.2
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
